move DeveloperProfileActionComponent to Developers namespace
