--- app/components/developer_profile_action_component.rb
@@ -1,11 +1,0 @@
-class DeveloperProfileActionComponent < ApplicationComponent
-  attr_reader :user
-
-  def initialize(user)
-    @user = user
-  end
-
-  def update_profile?
-    user&.developer
-  end
-end
