+++ app/components/developers/profile_action_component.rb
@@ -1,0 +1,13 @@
+module Developers
+  class ProfileActionComponent < ApplicationComponent
+    attr_reader :user
+
+    def initialize(user)
+      @user = user
+    end
+
+    def update_profile?
+      user&.developer
+    end
+  end
+end
