+++ config/locales/en.yml
@@ -169,14 +169,6 @@ en:
   developer_mailer:
     first_message:
       subject: Tips on responding to your first message on railsdevs
-  developer_profile_action_component:
-    add_profile: Add my profile
-    edit_profile: Update my profile
-  developer_query_sort_button_component:
-    sort:
-      availability: Availability
-      newest: Newest
-      title: Sort
   developers:
     admin_component:
       copy: Copy
@@ -248,6 +240,9 @@ en:
     new_fields_component:
       cta: Complete my profile
       title: New fields were added to developer profiles!
+    profile_action_component:
+      add_profile: Add my profile
+      edit_profile: Update my profile
     query_component:
       apply: Apply
       clear: Clear
@@ -263,6 +258,11 @@ en:
       title: Developer filters
       work_preference:
         title: Work preference
+    query_sort_button_component:
+      sort:
+        availability: Availability
+        newest: Newest
+        title: Sort
     search_status:
       actively_looking: Actively looking for work
       not_interested: Not interested in work right now
