+++ config/locales/pt-BR.yml
@@ -116,9 +116,6 @@ pt-BR:
   date:
     formats:
       month_year: "%B / %Y"
-  developer_profile_action_component:
-    add_profile: Adicionar meu perfil
-    edit_profile: Atualizar meu perfil
   developers:
     availability_component:
       in_future_html: Disponível em <time datetime="%{date}">%{date_in_words}</time>
@@ -171,6 +168,9 @@ pt-BR:
     index:
       description: "%{developers_count}+ desenvolvedores Ruby on Rails buscando seu próximo trabalho. Aqui você vai encontrar de Juniors a seniors e tudo no meio do caminho."
       title: Desenvolvedores Ruby on Rails
+    profile_action_component:
+      add_profile: Adicionar meu perfil
+      edit_profile: Atualizar meu perfil
     query_component:
       apply: Aplicar
       close_menu: Fechar menu
